stego_backend/stego/img_stego/requirements.txt
albumentations==0.4.3
bchlib==0.14.0
einops==0.3.0
imagenet_c==0.0.3
kornia==0.6.0
lmdb==1.4.1
lpips==0.1.4
matplotlib==3.4.2
numpy==1.24.4
omegaconf==2.1.1
open_clip_torch==2.0.2
opencv_python==3.4.18.65
opencv_python_headless==4.8.1.78
pandas==1.4.2
Pillow==9.4.0
Pillow==10.3.0
pytorch_lightning==1.4.2
reedsolo==1.7.0
Requests==2.31.0
safetensors==0.4.3
scikit_image==0.19.2
scipy==1.8.0
six==1.16.0
slackclient==2.9.3
taming_transformers==0.0.1
timm==0.9.16
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.19.2
xformers==0.0.26.post1
